Added auth to journal entries and configured spring security

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -7,6 +7,7 @@ import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -22,15 +23,17 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+    @Transactional
     public void saveEntry(JournalEntry journalEntry, String username) {
         try {
             User user = userService.findByUsername(username);
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved); // After saving journal entry inside journal_entries, we add it to the arraylist created inside User
-            userService.saveEntry(user); // we are saving the user now into the users collection
+            userService.saveUser(user); // we are saving the user now into the users collection
         } catch (Exception e) {
             log.error("Exception: ", e);
+            throw e;
         }
     }
 
@@ -46,15 +49,22 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteJournalEntryById(String username, ObjectId id) {
+    @Transactional
+    public boolean deleteJournalEntryById(String username, ObjectId id) {
         User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed) {
+            userService.saveNewUser(user);
+            journalEntryRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public JournalEntry updateJournalEntry(JournalEntry journalEntry) {
         return journalEntryRepository.save(journalEntry);
     }
+
+
 
 }
